feat(image): allow async `getSrcToShow()` in widget
This means it’s possible to do stuff like fetch image contents
and supply base64 representation of it, instead of having to provide
a src that points somewhere.

This works OK in authoring widget environment,
but not in DOM serialization environment where async is not an option
and actual location should be passed synchronuously.